+++ src/values/ListValue.ts
@@ -180,4 +180,16 @@ export default class ListValue extends SimpleValue {
         for (const value of this.values) sum += value.getSize();
         return sum;
     }
+
+    isCollection() {
+        return true;
+    }
+
+    getTranslationItems(): Value[] {
+        return this.values;
+    }
+
+    createTranslation(creator: Expression, results: Value[]): Value {
+        return new ListValue(creator, results);
+    }
 }